refactor: splitting code into files
