+++ src/types/types.ts
@@ -11,4 +11,5 @@ export interface IProduct {
   thumbnail: string;
   images: string[];
   orderNumber?: number;
+  count?: number;
 }
